Charge species must be a float
Must be merged after #248.
Closes #260.

See merge request gysela-developpers/gyselalibxx!573

--------------------------------------------

+++ src/geometryXVx/poisson/chargedensitycalculator.cpp
@@ -16,7 +16,7 @@ DSpanX ChargeDensityCalculator::operator()(DSpanX const rho, DViewSpXVx const al
     IndexSp const last_species = ddc::discrete_space<IDimSp>().charges().domain().back();
     double chargedens_adiabspecies = 0.;
     if (last_kin_species != last_species) {
-        chargedens_adiabspecies = double(charge(last_species));
+        chargedens_adiabspecies = charge(last_species);
     }
 
     // reduction over species and velocity space
@@ -24,13 +24,13 @@ DSpanX ChargeDensityCalculator::operator()(DSpanX const rho, DViewSpXVx const al
             = allfdistribu.allocation_kokkos_view();
     Kokkos::View<double*, Kokkos::LayoutRight> const rho_view = rho.allocation_kokkos_view();
 
-    host_t<ViewSp<int>> const charges = ddc::host_discrete_space<IDimSp>().charges();
-    host_t<ViewSp<int>> const kinetic_charges = charges[allfdistribu.domain<IDimSp>()];
+    host_t<ViewSp<double>> const charges = ddc::host_discrete_space<IDimSp>().charges();
+    host_t<ViewSp<double>> const kinetic_charges = charges[allfdistribu.domain<IDimSp>()];
 
     auto charges_alloc
             = create_mirror_view_and_copy(Kokkos::DefaultExecutionSpace(), kinetic_charges);
 
-    Kokkos::View<const int*, Kokkos::LayoutRight> const charges_view
+    Kokkos::View<const double*, Kokkos::LayoutRight> const charges_view
             = charges_alloc.span_cview().allocation_kokkos_view();
 
     Kokkos::View<const double*, Kokkos::LayoutRight> const coef_view
@@ -50,7 +50,7 @@ DSpanX ChargeDensityCalculator::operator()(DSpanX const rho, DViewSpXVx const al
                         [&](int const& ivx, double& sum) {
                             // [TO DO] Nested reduction may be possible?
                             for (std::size_t isp = 0; isp < nsp; isp++) {
-                                sum += static_cast<double>(charges_view(isp)) * coef_view(ivx)
+                                sum += charges_view(isp) * coef_view(ivx)
                                        * allfdistribu_view(isp, ix, ivx);
                             }
                         },
